merged the changes to StructTree.java and ParseContentStream.java from check_metadata branch.
Here's the latest commit from that branch:
now accounts for the DP operator which also marks the beginning of marked content, and
has a dictionary with a required MCID key as its operator. now stores two sets where one holds all
the mcids for the page and the other holds the ones specifically for images. also accounts for a very
basic way of knowing if a graph or table is being drawn using the "l" operator, which is short for lineto
in PostScript and what it does is "Append straight line segment to path."